Feat: Route security

+++ server/api/admin.js
@@ -5,8 +5,10 @@ module.exports = router
 //GET: /admin/users
 router.get('/users', async (req, res, next) => {
   try {
-    const users = await User.findAll()
-    res.json(users)
+    if (req.user && req.user.dataValues && req.user.dataValues.admin) {
+      const users = await User.findAll()
+      res.json(users)
+    } else res.send('Admin access only.')
   } catch (err) {
     next(err)
   }
@@ -14,12 +16,14 @@ router.get('/users', async (req, res, next) => {
 
 router.delete('/users/:id', async (req, res, next) => {
   try {
-    await User.destroy({
-      where: {
-        id: req.params.id
-      }
-    })
-    res.sendStatus(200)
+    if (req.user && req.user.dataValues && req.user.dataValues.admin) {
+      await User.destroy({
+        where: {
+          id: req.params.id,
+        },
+      })
+      res.sendStatus(200)
+    } else res.send('Admin access only.')
   } catch (err) {
     next(err)
   }
@@ -27,8 +31,10 @@ router.delete('/users/:id', async (req, res, next) => {
 
 router.get('/candy', async (req, res, next) => {
   try {
-    const candies = await Candy.findAll()
-    res.json(candies)
+    if (req.user && req.user.dataValues && req.user.dataValues.admin) {
+      const candies = await Candy.findAll()
+      res.json(candies)
+    } else res.send('Admin access only.')
   } catch (err) {
     next(err)
   }
